trading/requirements.txt
mysql-connector-python
alpaca-trade-api
lumibot